+++ docs/changes.rst
@@ -91,6 +91,15 @@ Released: not yet
 * Improved the unit test cases for the `Client` class and for the zhmcclient
   exception classes, and migrated them to py.test.
 
+* Migrated the unit tests for HBAs from unittest to py.test, and
+  improved the test cases.
+
+* In the `Hba.reassign_port()` method, updated the `Hba` object with the
+  changed port, consistent with other update situations.
+
+* Clarified in the description of `HbaManager.list()` that only the
+  'element-uri' property is returned and can be used for filtering.
+  
 **Known issues:**
 
 * See `list of open issues`_.